# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API ",
        "description": "Listado de los productos extras",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000"
        }
    ],
    "paths": {
...
        "/api/Extras/{id}": {
            "get": {
                "tags": [
                    "Extras"
                ],
                "summary": "Obtener un Extra específico",
                "description": "Retorna los detalles de un Extra con el ID proporcionado.",
                "operationId": "7519a5fd45ea95c9c2d2ceb81aab8868",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "Logotipos": {
                                            "type": "string",
                                            "example": "Logo1"
                                        },
                                        "Imagenes editadas": {
                                            "type": "string",
                                            "example": "img1"
                                        },
                                        "Cantidad": {
                                            "type": "integer",
                                            "example": 2
                                        },
                                        "Fecha de entrega": {
                                            "type": "string",
                                            "format": "date-time",
                                            "example": "2023-02-23T12:33:45.000000Z"
                                        },
                                        "Precio": {
                                            "type": "number",
                                            "format": "float",
                                            "example": 10.5
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "No se encontró el Extra",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "No query results for model [App\\\\Models\\\\Extras] #id"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Extras"
                ],
                "summary": "Actualizar un Extra existente",
                "description": "Actualiza los detalles de un Extra con el ID proporcionado.",
                "operationId": "cc317ed9ebc35cbae075c015f3efaf95",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "Logotipos",
                                    "Imagenes editadas",
                                    "Cantidad",
                                    "Fecha de entrega",
                                    "Precio"
                                ],
                                "properties": {
                                    "Logotipos": {
                                        "type": "string",
                                        "example": "Logo1"
                                    },
                                    "Imagenes editadas": {
                                        "type": "string",
                                        "example": "img1"
                                    },
                                    "Cantidad": {
                                        "type": "integer",
                                        "example": 3
                                    },
                                    "Fecha de entrega": {
                                        "type": "string",
                                        "format": "date-time",
                                        "example": "2023-02-23T12:33:45.000000Z"
                                    },
                                    "Precio": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 15.5
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Extra actualizado correctamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "Logotipos": {
                                            "type": "string",
                                            "example": "Logo1"
                                        },
                                        "Imagenes editadas": {
                                            "type": "string",
                                            "example": "img1"
                                        },
                                        "Cantidad": {
                                            "type": "integer",
                                            "example": 3
                                        },
                                        "Fecha de entrega": {
                                            "type": "string",
                                            "format": "date-time",
                                            "example": "2023-02-23T12:33:45.000000Z"
                                        },
                                        "Precio": {
                                            "type": "number",
                                            "format": "float",
                                            "example": 15.5
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "No se encontró el Extra",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "No query results for model [App\\\\Models\\\\Extras] #id"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Extras"
                ],
                "summary": "Eliminar un Extra",
                "description": "Elimina un Extra del sistema usando el ID proporcionado.",
                "operationId": "4fab8dda985ec4b1b3f7ed8898394df2",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Extra eliminado correctamente"
                    },
                    "404": {
                        "description": "No se encontró el Extra",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "No query results for model [App\\\\Models\\\\Extras] #id"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
...
    "components": {
        "securitySchemes": {
            "bearer_token": {
                "type": "apiKey",
                "description": "Sanctum bearer token",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
    "security": [
        {
            "bearer_token": []
        }
    ]
}